Update the title and description of the introduction screen
