+++ app/constants/component_constant.js
@@ -7,3 +7,4 @@ export const headerWithAudioMaxHeight = getStyleOfDevice(300, 240);
 export const headerWithAudioMinHeight = 170;
 export const headerWithAudioScrollDistance = (headerWithAudioMaxHeight - headerWithAudioMinHeight);
 export const screenPaddingHorizontal = 16;
+export const introViewPaddingHorizontal = 24;
