+++ app/constants/intro_constant.js
@@ -1,22 +1,30 @@
 export const slides = [
   {
     key: 'one',
-    title: 'Introductin title 1',
-    text: 'Description of the introduction screen 1.\nContent of the indroduction 1',
+    title: 'ការប្រែប្រួលរូបរាង្គកាយពេលពេញវ័យ',
+    text: 'យល់ដឹងពីចំនុចសំខាន់ៗ និងការត្រៀមខ្លួននានានៅពេលអ្នកពេញវ័យ និងកត្តាផ្សេងៗដែលជិះឥទ្ធិពលមិនល្អដល់ខ្លួនអ្នក',
+    // text: 'Description of the introduction screen 1.\nContent of the indroduction 1',
     image: require('../assets/images/intro_2.jpg'),
     backgroundColor: '#59b2ab',
   },
   {
     key: 'two',
-    title: 'Introduction title 2',
-    text: 'Description of the introduction screen 2.\nContent of the indroduction 2',
+    title: 'យល់ដឹងពីសុខភាពផ្លូវភេទ',
+    text: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស',
     image: require('../assets/images/intro_2.jpg'),
     backgroundColor: '#febe29',
   },
   {
     key: 'three',
-    title: 'Introuction title 3',
-    text: 'Description of the introduction screen 3.\nContent of the indroduction 3',
+    title: 'យល់ដឹងអំពីការមានផ្ទៃពោះ',
+    text: 'ការប្រឈមមុខមានផ្ទៃពោះដោយឥតព្រាងទុក ជាមួយសេវាប្រឹេក្សា និងការរក្សាការសំងាត់',
+    image: require('../assets/images/intro_3.jpg'),
+    backgroundColor: '#22bcb5',
+  },
+  {
+    key: 'four',
+    title: 'កន្លែងផ្តល់សេវានិងព្រឹក្សាយោបល់',
+    text: 'អ្នកអាចរកមេីលគ្លីនិកដែលនៅជិតលោកអ្នកជាមួយព័តិមានលំអិតអំពី សេវាកម្ម  ម៉ោងផ្តល់សេវា និងផែនទីងាយស្រួលក្នុងការរកទីតាំងគ្លីនិចនិមួយៗ',
     image: require('../assets/images/intro_3.jpg'),
     backgroundColor: '#22bcb5',
   }
