+++ app/views/main/MainView.js
@@ -8,7 +8,8 @@ import navigatorUtil from '../../utils/navigator_util';
 const MainView = () => {
   useEffect(() => {
     const initialNavigation = async () => {
-      navigationRef.current?.reset({ index: 0, routes: [{ name: await navigatorUtil.getInitialRouteName() }] });
+      navigationRef.current?.reset({ index: 0, routes: [{ name: 'IntroductionView' }] });
+      // navigationRef.current?.reset({ index: 0, routes: [{ name: await navigatorUtil.getInitialRouteName() }] });
       appStatusService.handleAppLaunchingStatus();
     }
